develop

+++ src/main/java/com/travelGraph/services/UserService.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 import com.travelGraph.dto.auth.CurrentUserDTO;
 import com.travelGraph.dto.user.CreateUserRequestDTO;
 import com.travelGraph.dto.user.UpdateUserRequestDTO;
-import com.travelGraph.entities.UserEntity;
+import com.travelGraph.entities.UserNode;
 import com.travelGraph.helpers.UpdateValueHelper;
 import com.travelGraph.repositories.UserRepository;
 import com.travelGraph.services.exceptions.DatabaseException;
@@ -27,52 +27,63 @@ public class UserService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    public List<UserEntity> findAll() {
+    public List<UserNode> findAll() {
         return userRepository.findAll();
     }
 
-    public UserEntity findById(Long id) {
-        Optional<UserEntity> user = userRepository.findById(id);
-        return user.orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
-    }
-
-    public Optional<UserEntity> findByEmail(String email) {
-        Optional<UserEntity> user = userRepository.findByEmail(email);
+    public Optional<UserNode> findById(String id) {
+        Optional<UserNode> user = userRepository.findById(id);
         return user;
     }
 
-    public void delete(CurrentUserDTO currentUser, Long id) {
+    public Optional<UserNode> findByEmail(String email) {
+        Optional<UserNode> user = userRepository.findByEmail(email);
+        return user;
+    }
+
+    public void delete(CurrentUserDTO currentUser, String id) {
         try {
             if (currentUser.getId().equals(id)) {
                 throw new DatabaseException("Usuário não pode deletar a si mesmo");
             }
 
-            findById(id);
+            Optional<UserNode> user = findById(id);
+
+            if (user.isEmpty()) {
+                throw new ResourceNotFoundException("Usuário não encontrado");
+            }
+
             userRepository.deleteById(id);
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException(e.getMessage());
         }
     }
 
-    public UserEntity update(Long id, UpdateUserRequestDTO updateUser) {
-        UserEntity entity = findById(id);
+    public UserNode update(String id, UpdateUserRequestDTO updateUser) {
+        Optional<UserNode> entity = findById(id);
 
-        String passwordEncryption = passwordEncoder.encode(updateUser.getPassword());
-        updateUser.setPassword(passwordEncryption);
-        
-        updateData(entity, updateUser);
+        if (entity.isEmpty()) {
+            throw new ResourceNotFoundException("Usuário não encontrado");
+        }
 
-        return userRepository.save(entity);
+        if (updateUser.getPassword() != null) {
+            String passwordEncryption = passwordEncoder.encode(updateUser.getPassword());
+            updateUser.setPassword(passwordEncryption);
+        }
+
+        updateData(entity.get(), updateUser);
+
+        return userRepository.save(entity.get());
     }
 
-    private void updateData(UserEntity entity, UpdateUserRequestDTO updateUser) {
+    private void updateData(UserNode entity, UpdateUserRequestDTO updateUser) {
         UpdateValueHelper.updateIfNotNull(entity::setName, updateUser.getName());
         UpdateValueHelper.updateIfNotNull(entity::setEmail, updateUser.getEmail());
         UpdateValueHelper.updateIfNotNull(entity::setPassword, updateUser.getPassword());
     }
 
-    public UserEntity create(@Valid CreateUserRequestDTO createUser) {
-        Optional<UserEntity> findUserByEmail = userRepository.findByEmail(createUser.getEmail());
+    public UserNode create(@Valid CreateUserRequestDTO createUser) {
+        Optional<UserNode> findUserByEmail = userRepository.findByEmail(createUser.getEmail());
 
         if (findUserByEmail.isPresent()) {
             throw new UnprocessableEntityException("Já existe um usuário com esse e-mail");
@@ -80,7 +91,7 @@ public class UserService {
         
         String passwordEncryption = passwordEncoder.encode(createUser.getPassword());
 
-        UserEntity user = new UserEntity();
+        UserNode user = new UserNode();
         user.setEmail(createUser.getEmail());
         user.setName(createUser.getName());
         user.setPassword(passwordEncryption);
